# apps/taptrade-platform/go-platform/services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tap Trade API
  version: "1.0.0"
  description: >
    Tap Trade prediction-market gateway API for non-redeemable gameplay points.
    Binary YES/NO outcomes use point-cents subunits on a 0-100 implied
    probability scale; settled winning shares return 100 point-cents per share.
    Points are gameplay-only units with no external value path.

    Coverage: this spec documents the core public, trading, portfolio,
    reward, loyalty, social, leaderboard, bot, auth proxy,
    admin account-review, admin discovery taxonomy, admin market operations,
    and admin market-lifecycle surfaces verified against the live handlers and
    DTOs as of 2026-06. CMS and remaining compatibility surfaces are
    intentionally out of scope here and tracked for documentation in
    IMPROVEMENT_PLAN.md P2-05 / G-04. The previous inherited spec is archived
    outside this launch document.
servers:
  - url: http://localhost:18080
    description: Local development
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly session cookie issued by the auth service. Mutating requests also require the X-CSRF-Token header (double-submit cookie).
    BotApiKey:
      type: http
      scheme: bearer
      description: "Bot/partner API key: Authorization: Bearer tna_<prefix>_<secret>. Scopes: read, trade."
  schemas:
...
    AdminPageMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        hasNext: { type: boolean }
    AdminPunter:
      type: object
      description: Admin account identity row for prediction-platform review.
      properties:
        id: { type: string }
        email: { type: string }
        username: { type: string, nullable: true }
        status: { type: string, enum: [active, suspended, deactivated] }
        countryCode: { type: string, nullable: true }
        createdAt: { type: string }
        lastLoginAt: { type: string, nullable: true }
      required: [id, email, status]
    AdminPunterListItem:
      allOf:
        - { $ref: "#/components/schemas/AdminPunter" }
        - type: object
          description: Admin account row enriched with point-account summary fields.
          properties:
            pointAccountBalancePoints: { type: integer, format: int64, description: Preferred available point-account subunits for review. }
            realizedPoints: { type: integer, format: int64, description: Preferred settled prediction result subunits. }
            unit: { type: string, enum: [PTS] }
...
    AdminAuditLog:
      type: object
      description: Admin audit row with actor, action, target, status, and details metadata.
      properties:
        id: { type: string }
        actorId: { type: string, nullable: true }
        action: { type: string }
        resourceType: { type: string, nullable: true }
        targetId: { type: string, nullable: true }
        status: { type: string }
        details: { type: object, additionalProperties: true }
        occurredAt: { type: string }
...
paths:
...
  /api/v1/admin/punters:
    get:
      summary: Admin list prediction accounts
      description: Lists prediction-platform accounts with point-account and settled-result summary fields. This read-only review surface does not move points.
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [active, suspended, deactivated] } }
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200":
          description: Account review rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AdminPunterListItem" }
                  pagination: { $ref: "#/components/schemas/AdminPageMeta" }
        "403": { description: Admin role required }
...
  /api/v1/admin/audit-logs:
    get:
      summary: Admin audit-log review
      description: Lists admin audit rows and in-process operation audit metadata with pagination and filters.
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: action, in: query, schema: { type: string } }
        - { name: resourceType, in: query, schema: { type: string } }
        - { name: actorId, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200":
          description: Audit rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AdminAuditLog" }
                  pagination: { $ref: "#/components/schemas/AdminPageMeta" }
...